refactor: move template functions to ccbr_tools.templates.__init__
