+++ tests/test_pkg_util.py
@@ -1,11 +1,5 @@
-from ccbr_tools.pkg_util import read_template
+from ccbr_tools.pkg_util import repo_base
 
 
-def test_read_template():
-    template_str = read_template("submit_slurm.sh")
-    assert all(
-        [
-            template_str.startswith("#!/usr/bin/env bash"),
-            template_str.endswith("{RUN_COMMAND}\n"),
-        ]
-    )
+def test_repo_base():
+    assert str(repo_base()).endswith("ccbr_tools")
